--- src/pages/mpl-404/getting-started/index.md
@@ -1,15 +1,0 @@
----
-title: Getting Started
-metaTitle: MPL-404 - Getting Started
-description: Get started with MPL-404 Smart Escrow
----
-
-Select the language / library you want to use below to get started with the new Metaplex Core Standard.
-
-{% quick-links %}
-
-{% quick-link title="JavaScript" icon="JavaScript" href="/mpl-404/getting-started/js" description="Get started with our JavaScript library based on the Umi framework." /%}
-
-{% quick-link title="Rust" icon="Rust" href="/core/getting-started/rust" description="Get started using our Rust crate." /%}
-
-{% /quick-links %}
