--- src/pages/mpl-404/getting-started/rust.md
@@ -1,7 +1,0 @@
----
-title: Getting Started using Rust
-metaTitle: MPL-404 - Getting Started - Rust
-description: Get started with MPL-404 using Rust
----
-
-Metaplex provides a Rust library that can be used to define hybrid assets.
