--- src/pages/mpl-404/index.md
@@ -1,19 +1,0 @@
----
-title: Overview
-metaTitle: MPL-404 - Overview
-description: Provides a high-level overview of the framework and on-chain protocol for hybrid assets.
----
-
-MPL-404 is a new model for digital assets, web3 games, and on-chain communities. At the core of the model is a swap program that trades a fixed number of fungible assets for a non-fungible asset and vice versa. The swap is a dual escrow system, ensuring that all available non-fungible assets are backed by escrowed fungibles and vice versa.
-
-## Swapping
-
-The ability to freely move between fungible and non-fungible assets in a predictable way allows this new asset class (often called 404s) to take advantage of the best aspects of both asset types. Non-fungible assets can now benefit from the liquidity, distribution, and DeFi opportunities associated with fungible assets. Conversely, fungible assets can now benefit from the improved utility, collectability, and identity that come from the non-fungible world.
-
-## Re-Rolling
-
-MPL-404 includes the option to "re-roll" the asset each time its swapped. For example, every non-fungible asset can have its metadata blanked as it enters the escrow wallet and randomly reassigned (rerolled) as it leaves escrow. The creator has the ability to both manage available traits as well as to charge a small fee during the swap (typically when swapping into an NFT). MPL-404 can add add “loot box” gamification to every 404 project and can serve as an alternative source of revenue (e.g. unique limited time traits for an NFT community or randomized in-game rewards). It also offers the ability to craft more dynamic collections that can evolve to better suit the needs of the project and community over time.
-
-{% callout type="note" %}
-MPL-404 is currently in alpha with fully self-serve docs and SDKs coming soon. Interested in using MPL-404 for your project or learning more? Contact us in [Discord](https://discord.gg/metaplex) and tell us more about what you're working on.
-{% /callout %}
