+++ src/pages/mpl-hybrid/getting-started/index.md
@@ -1,0 +1,13 @@
+---
+title: Getting Started
+metaTitle: MPL-Hybrid - Getting Started
+description: Get started with the MPL-Hybrid Smart Escrow program.
+---
+
+Select the language / library you want to use below to get started with the new Metaplex Core Standard.
+
+{% quick-links %}
+
+{% quick-link title="JavaScript" icon="JavaScript" href="/mpl-hybrid/getting-started/js" description="Get started with our JavaScript library based on the Umi framework." /%}
+
+{% /quick-links %}
