+++ src/pages/mpl-hybrid/initializeNFTData.md
