+++ src/pages/mpl-hybrid/philosophy.md
